⚡️ ✨ More example plots and isort. Metric comupation optimization
Added the thomas attractor and one single example plot only.

Restructured code when standardization is done we can use all data anyway and dont have to do it on the subsequence. saves computation power. Also the curvature slice is better since the beginning and the end especially of short subsequences are slightly different if curvature is only calculated on the subsequence and not as whole sequence.